@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class HarryKartApp {
-    public static void main(String ... args) {
+    public static void main(String... args) {
         SpringApplication.run(HarryKartApp.class, args);
     }
 }
