@@ -1,16 +1,13 @@
 package se.atg.service.harrykart.java.model;
 
-import java.util.Objects;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class Loop {
-
-    public final int number;
-
+public record Loop(int number) {
     public static Loop of(int number) {
         return new Loop(number);
     }
 
-    private Loop(int number) {
+    public Loop(@JsonProperty("number") int number) {
         this.number = number;
     }
 
@@ -19,18 +16,5 @@ public class Loop {
         return "Loop{" +
                 "number=" + number +
                 '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Loop loop = (Loop) o;
-        return number == loop.number;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(number);
     }
 }
