@@ -1,15 +1,14 @@
 package se.atg.service.harrykart.java.model;
 
-public class PowerUp {
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-    public final int value;
+public record PowerUp(int value) {
 
     public static PowerUp of(int value) {
         return new PowerUp(value);
     }
 
-    private PowerUp(int value) {
+    public PowerUp(@JsonProperty("value") int value) {
         this.value = value;
     }
-
 }
